# backend/src/main/java/com/sandsational/backend/dto/request/LoginUserRequest.java
package com.sandsational.backend.dto.request;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record LoginUserRequest (
    @NotBlank
    @Email
    @Size(max=254)
    String email,

    @NotBlank
    @Size(min=8, max=72)
    String password
){

}