# backend/src/main/java/com/sandsational/backend/dto/response/ErrorResponse.java
package com.sandsational.backend.dto.response;

import java.time.Instant;

public record ErrorResponse(
        Instant timestamp,
        int status,
        String error,
        String message,
        String path
) {
}